Add files via upload
This project seamlessly integrates Angular for the frontend and Django for the backend to create a dynamic and responsive web application. The backend is powered by Django models connected to PostgreSQL for efficient data management. Secure REST APIs enable real-time GET and POST operations, while Angular components dynamically display and interact with data.